# [Library] Fix legacy systolic array performance issue