--- a/allo/library/systolic.py
+++ b/allo/library/systolic.py
@@ -294,7 +294,6 @@ def schedule_systolic(s):
         assert len(s.inst_list) == 9
         tile_name = "systolic_tile"
         M0, M1 = s.inst_list[-3], s.inst_list[-2]
-        kernel_loop = s.get_loops(f"PE_kernel_packed_int8xint8")["reduction"]["k"]
     elif s.top_func_name == "packed_int8xint8_systolic":
         assert len(s.inst_list) == 6
         tile_name = "systolic_tile"
@@ -317,7 +316,15 @@ def schedule_systolic(s):
     outer_tile_loop = s.get_loops(s.top_func_name)["outer_tile"]["mi"]
     tile_loop = s.fuse(outer_tile_loop, inner_tile_loop)
     s.dataflow(tile_loop)
-    s.pipeline(kernel_loop)
+    kernel_loop = None
+    for kernel_name in {"PE_kernel", "PE_kernel_packed_int8xint8"}:
+        try:
+            kernel_loop = s.get_loops(kernel_name)["reduction"]["k"]
+            break
+        except RuntimeError:
+            continue
+    if kernel_loop is not None:
+        s.pipeline(kernel_loop)
     pe = s.unfold(f"{tile_name}:PE", [0, 1])  # specify which are spatial loops
     s.to(MockBuffer(tile_name, "A_fifo"), pe, axis=1, depth=M0 + 1)
     s.to(MockBuffer(tile_name, "B_fifo"), pe, axis=0, depth=M1 + 1)