<!--- Copyright Allo authors. All Rights Reserved. -->
<!--- SPDX-License-Identifier: Apache-2.0  -->

## Description ##
This PR fixes a legacy SA issue where the original SA based on Allo's schedule cannot achieve the expected performance.
This is mainly due to the incorrect pipelined loops for data loading/storing.

### Problems ###
The old performance result (8x8 GEMM using 2x2 SA):
```
+ Latency: 
    * Summary: 
    +---------+---------+----------+----------+-----+-----+----------+
    |  Latency (cycles) |  Latency (absolute) |  Interval | Pipeline |
    |   min   |   max   |    min   |    max   | min | max |   Type   |
    +---------+---------+----------+----------+-----+-----+----------+
    |      725|      727|  2.416 us|  2.423 us|  454|  454|  dataflow|
    +---------+---------+----------+----------+-----+-----+----------+

    + Detail: 
        * Instance: 
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+
        |                 |              |  Latency (cycles) |  Latency (absolute) |  Interval |                 Pipeline                 |
        |     Instance    |    Module    |   min   |   max   |    min   |    max   | min | max |                   Type                   |
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+
        |entry_proc_U0    |entry_proc    |        0|        0|      0 ns|      0 ns|    0|    0|                                        no|
        |load_buf0_1_U0   |load_buf0_1   |      136|      137|  0.453 us|  0.457 us|   64|   64|  loop rewind stp(delay=0 clock cycles(s))|
        |load_buf1_1_U0   |load_buf1_1   |      136|      137|  0.453 us|  0.457 us|   64|   64|  loop rewind stp(delay=0 clock cycles(s))|
        |systolic_U0      |systolic      |      453|      453|  1.510 us|  1.510 us|  453|  453|                                        no|
        |store_res2_1_U0  |store_res2_1  |      134|      135|  0.447 us|  0.450 us|   64|   64|  loop rewind stp(delay=0 clock cycles(s))|
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+

+ Latency: 
    * Summary: 
    +---------+---------+-----------+-----------+-----+-----+----------+
    |  Latency (cycles) |   Latency (absolute)  |  Interval | Pipeline |
    |   min   |   max   |    min    |    max    | min | max |   Type   |
    +---------+---------+-----------+-----------+-----+-----+----------+
    |       16|       16|  53.328 ns|  53.328 ns|   14|   14|  dataflow|
    +---------+---------+-----------+-----------+-----+-----+----------+

    + Detail: 
        * Instance: 
        +-----------------------------------------------+--------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+
        |                                               |                                            |  Latency (cycles) |   Latency (absolute)  |  Interval | Pipeline|
        |                    Instance                   |                   Module                   |   min   |   max   |    min    |    max    | min | max |   Type  |
        +-----------------------------------------------+--------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+
        |systolic_tile_Loop_l_data_load_k4_proc411_U0   |systolic_tile_Loop_l_data_load_k4_proc411   |       10|       10|  33.330 ns|  33.330 ns|   10|   10|       no|
        |PE_kernel_0_012_U0                             |PE_kernel_0_012                             |       13|       13|  43.329 ns|  43.329 ns|   13|   13|       no|
        |PE_kernel_1_013_U0                             |PE_kernel_1_013                             |       13|       13|  43.329 ns|  43.329 ns|   13|   13|       no|
        |PE_kernel_0_114_U0                             |PE_kernel_0_114                             |       13|       13|  43.329 ns|  43.329 ns|   13|   13|       no|
        |PE_kernel_1_115_U0                             |PE_kernel_1_115                             |       13|       13|  43.329 ns|  43.329 ns|   13|   13|       no|
        |systolic_tile_Loop_l_data_drain_k5_proc516_U0  |systolic_tile_Loop_l_data_drain_k5_proc516  |       10|       10|  33.330 ns|  33.330 ns|   10|   10|       no|
        +-----------------------------------------------+--------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+

```

The new performance result is shown below:
```
+ Latency: 
    * Summary: 
    +---------+---------+----------+----------+-----+-----+----------+
    |  Latency (cycles) |  Latency (absolute) |  Interval | Pipeline |
    |   min   |   max   |    min   |    max   | min | max |   Type   |
    +---------+---------+----------+----------+-----+-----+----------+
    |      397|      399|  1.322 us|  1.329 us|  252|  252|  dataflow|
    +---------+---------+----------+----------+-----+-----+----------+

    + Detail: 
        * Instance: 
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+
        |                 |              |  Latency (cycles) |  Latency (absolute) |  Interval |                 Pipeline                 |
        |     Instance    |    Module    |   min   |   max   |    min   |    max   | min | max |                   Type                   |
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+
        |entry_proc_U0    |entry_proc    |        0|        0|      0 ns|      0 ns|    0|    0|                                        no|
        |load_buf0_1_U0   |load_buf0_1   |       73|       74|  0.243 us|  0.246 us|   64|   64|  loop rewind stp(delay=0 clock cycles(s))|
        |load_buf1_1_U0   |load_buf1_1   |       73|       74|  0.243 us|  0.246 us|   64|   64|  loop rewind stp(delay=0 clock cycles(s))|
        |systolic_U0      |systolic      |      251|      251|  0.836 us|  0.836 us|  251|  251|                                        no|
        |store_res2_1_U0  |store_res2_1  |       71|       72|  0.236 us|  0.240 us|   64|   64|  loop rewind stp(delay=1 clock cycles(s))|
        +-----------------+--------------+---------+---------+----------+----------+-----+-----+------------------------------------------+

+ Latency: 
    * Summary: 
    +---------+---------+-----------+-----------+-----+-----+----------+
    |  Latency (cycles) |   Latency (absolute)  |  Interval | Pipeline |
    |   min   |   max   |    min    |    max    | min | max |   Type   |
    +---------+---------+-----------+-----------+-----+-----+----------+
    |       16|       16|  53.280 ns|  53.280 ns|   14|   14|  dataflow|
    +---------+---------+-----------+-----------+-----+-----+----------+

    + Detail: 
        * Instance: 
        +---------------------------------------------+------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+
        |                                             |                                          |  Latency (cycles) |   Latency (absolute)  |  Interval | Pipeline|
        |                   Instance                  |                  Module                  |   min   |   max   |    min    |    max    | min | max |   Type  |
        +---------------------------------------------+------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+
        |systolic_tile_Loop_l_data_load_k4_proc4_U0   |systolic_tile_Loop_l_data_load_k4_proc4   |       10|       10|  33.300 ns|  33.300 ns|   10|   10|       no|
        |PE_kernel_0_0_U0                             |PE_kernel_0_0                             |       13|       13|  43.290 ns|  43.290 ns|   13|   13|       no|
        |PE_kernel_1_0_U0                             |PE_kernel_1_0                             |       13|       13|  43.290 ns|  43.290 ns|   13|   13|       no|
        |PE_kernel_0_1_U0                             |PE_kernel_0_1                             |       13|       13|  43.290 ns|  43.290 ns|   13|   13|       no|
        |PE_kernel_1_1_U0                             |PE_kernel_1_1                             |       13|       13|  43.290 ns|  43.290 ns|   13|   13|       no|
        |systolic_tile_Loop_l_data_drain_k5_proc5_U0  |systolic_tile_Loop_l_data_drain_k5_proc5  |       10|       10|  33.300 ns|  33.300 ns|   10|   10|       no|
        +---------------------------------------------+------------------------------------------+---------+---------+-----------+-----------+-----+-----+---------+
```



## Checklist ##

Please make sure to review and check all of these items:
- [x] PR's title starts with a category (e.g. [Bugfix], [IR], [Builder], etc)
- [x] All changes have test coverage (It would be good to provide ~2 different test cases to test the robustness of your code)
- [x] Pass the [formatting check](https://cornell-zhang.github.io/allo/developer/index.html#id1) locally
- [x] Code is well-documented